+++ TM4C123_ACOMP/xDriver_MCU/WDT/App/Mode/xSource/WDT_ModeStruct.c
@@ -21,36 +21,36 @@
  * Date           Author     Version     Description
  * 24 jul. 2020     vyldram    1.0         initial Version@endverbatim
  */
-#include <stdlib.h>
 #include <xDriver_MCU/WDT/App/Mode/xHeader/WDT_ModeStruct.h>
+
+#include <stdlib.h>
 #include <xDriver_MCU/WDT/Peripheral/WDT_Peripheral.h>
 
 void WDT__vCreateModeStructPointer(WDT_nMODE enMode, WDT_MODE_Typedef* psMode)
 {
-
     if((uint32_t) 0U != (uint32_t) psMode )
     {
-        psMode->enInterrupt = (WDT_nINTERRUPT) ((uint32_t) enMode & 1U);
-        psMode->enStall = (WDT_nSTALL) (( (uint32_t) enMode >> 8U) & 1U );
-        psMode->enIntType = (WDT_nINT_TYPE) (( (uint32_t) enMode >> 16U ) & 1U);
-        psMode->enResetOutput = (WDT_nRESET) (( (uint32_t) enMode >> 24U ) & 1U);
+        psMode->enInterrupt = (WDT_nINTERRUPT) ((uint32_t) enMode & 1UL);
+        psMode->enStall = (WDT_nSTALL) (( (uint32_t) enMode >> 8UL) & 1UL);
+        psMode->enIntType = (WDT_nINT_TYPE) (( (uint32_t) enMode >> 16UL) & 1UL);
+        psMode->enResetOutput = (WDT_nRESET) (( (uint32_t) enMode >> 24UL) & 1UL);
     }
 }
 
 WDT_MODE_Typedef* WDT__psCreateModeStruct(WDT_nMODE enMode)
 {
-    WDT_MODE_Typedef* psMode = (WDT_MODE_Typedef*) 0U;
+    WDT_MODE_Typedef* psMode = (WDT_MODE_Typedef*) 0UL;
     #if defined (__TI_ARM__ )
-    psMode = (WDT_MODE_Typedef*) memalign((size_t) 4U, (size_t) sizeof(WDT_MODE_Typedef));
+    psMode = (WDT_MODE_Typedef*) memalign((size_t) 4UL, (size_t) sizeof(WDT_MODE_Typedef));
     #elif defined (__GNUC__ )
     psMode = (WDT_MODE_Typedef*) malloc((size_t) sizeof(WDT_MODE_Typedef));
     #endif
-    if((uint32_t) 0U != (uint32_t) psMode )
+    if(0UL != (uint32_t) psMode )
     {
-        psMode->enInterrupt = (WDT_nINTERRUPT) ( (uint32_t) enMode & 1U);
-        psMode->enStall = (WDT_nSTALL) (( (uint32_t) enMode >> 8U ) & 1U);
-        psMode->enIntType = (WDT_nINT_TYPE) (( (uint32_t) enMode >> 16U ) & 1U);
-        psMode->enResetOutput = (WDT_nRESET) (( (uint32_t) enMode >> 24U ) & 1U);
+        psMode->enInterrupt = (WDT_nINTERRUPT) ( (uint32_t) enMode & 1UL);
+        psMode->enStall = (WDT_nSTALL) (( (uint32_t) enMode >> 8UL) & 1UL);
+        psMode->enIntType = (WDT_nINT_TYPE) (( (uint32_t) enMode >> 16UL) & 1UL);
+        psMode->enResetOutput = (WDT_nRESET) (( (uint32_t) enMode >> 24UL) & 1UL);
     }
     return psMode;
 }
@@ -60,6 +60,3 @@ void WDT__vDeleteModeStruct(WDT_MODE_Typedef* psMode)
     free(psMode);
     psMode = (WDT_MODE_Typedef*) 0UL;
 }
-
-
-
